Add support for scanning.
Needed to add permissions for course and fine location as well as request
the permissions because I'm targeting SDK 23 (Android M).

One thing that's wrong is that when you have Bluetooth disabled you end
up with the UI for both enabling BT and the request to allow the app
the to use the location permissions.

I think its weird that location is required for BTLE scanning, but its
a requirement.